+++ packages/discovery/discover_routes.ts
@@ -2,11 +2,18 @@ import { fromFileUrl } from "@std/path/from-file-url";
 import { toFileUrl } from "@std/path/to-file-url";
 import { join } from "@std/path/join";
 import { parse } from "@std/path/parse";
-import { asSerializablePattern } from "./as_serializable_pattern.ts";
 import { asURLPatterns } from "@http/route/as-url-pattern";
-import type { PathPattern, RoutePattern } from "@http/route/types";
-
-type ParsedPath = ReturnType<typeof parse>;
+import type {
+  ComparableRoute,
+  DirectoryReader,
+  DiscoveredPath,
+  DiscoveredRoute,
+  PathMapper,
+  PathPattern,
+  RouteComparator,
+  RouteMapper,
+} from "./types.ts";
+import { isDirectory, isFile } from "./_dir_entry.ts";
 
 /**
  * Options for the `discoverRoutes` function
@@ -39,106 +46,52 @@ export interface DiscoverRoutesOptions {
    */
   consolidate?: boolean;
   /**
-   * Log all discovered routes.
+   * Log all discovered routes to console.
    */
   verbose?: boolean;
+  /**
+   * Function to return directory entries.
+   * Will default to `Deno.readDir` or Node's `opendir`.
+   */
+  readDir?: DirectoryReader;
 }
-
-/**
- * Function that maps a discovered path entry to zero, one or many routes.
- */
-export type RouteMapper = (
-  entry: DiscoveredPath,
-) => Iterable<DiscoveredRoute> | AsyncIterable<DiscoveredRoute>;
-
-/**
- * Function that transforms a discovered path entry.
- *
- * May be used to handle special case paths like 'index' files, and
- * transform one matching syntax (eg Fresh) to URLPattern syntax.
- */
-export type PathMapper = (
-  entry: DiscoveredPath,
-) => DiscoveredPath;
-
-/**
- * Compare two routes to determine ordering of the discovered routes.
- *
- * If `a` is a higher priority route than `b`, then this should return a negative number,
- * otherwise positive if lower, or zero if the priority is equal or indeterminate.
- *
- * Comparison maybe based on the pattern and/or the module of the route.
- */
-export type RouteComparator = (
-  a: ComparableRoute,
-  b: ComparableRoute,
-) => number;
-
-/**
- * A discovered route, that associates one or more routes to a module specifier.
- */
-export interface DiscoveredRoute {
-  pattern: RoutePattern;
-  module: string | URL;
-}
-
-/**
- * A simplified discovered route that associates a single URLPattern with a module specifier.
- */
-export interface ComparableRoute {
-  /**
-   * A single route pattern represented as a URLPattern.
-   */
-  pattern: URLPattern;
-  /**
-   * This will always be a URL object if it's an absolute URL,
-   * otherwise it will be a string if a relative URL or a bare specifier.
-   */
-  module: string | URL;
-}
-
-/**
- * A discovered file path, that may be a potential route.
- */
-export type DiscoveredPath = ParsedPath & {
-  /**
-   * The parent URL path pattern of this entry
-   */
-  // parentPattern: PathPattern;
-  /**
-   * The parent filesystem path of this entry
-   */
-  parentPath: string;
-  /**
-   * The default module URL path pattern (ie. parentPattern + filename without extension),
-   * the mapper fn may use this directly as the route, or provide an alternative.
-   */
-  pattern: PathPattern;
-  /**
-   * A default module URL for this entry (ie. file path url + filename with extension).
-   * NOTE: This may not actually be an importable module, if the file type itself cannot
-   * be imported by the JS engine. It's up to the mapper fn to determine this.
-   */
-  module: URL;
-};
 
 /**
  * Utility to walk the local filesystem and discover routes.
  *
+ * @example
+ * ```ts
+ * const routes = await discoverRoutes({
+ *   pattern: "/",
+ *   fileRootUrl: import.meta.resolve("./routes"),
+ *   verbose: true,
+ * });
+ * ```
+ *
+ * @param opts configuration options
  * @returns an array of discovered route tuples of the URLPattern to handler module URL.
  */
 export async function discoverRoutes(
   opts?: DiscoverRoutesOptions,
 ): Promise<DiscoveredRoute[]> {
-  const pathMapper: PathMapper = opts?.pathMapper ?? defaultPathMapper;
-  const routeMapper: RouteMapper = opts?.routeMapper ?? defaultRouteMapper;
-  const compare: RouteComparator = opts?.compare ?? defaultRouteCompare;
+  const pathMapper: PathMapper = opts?.pathMapper ??
+    (await import("./index_path_mapper.ts")).default;
+
+  const routeMapper: RouteMapper = opts?.routeMapper ??
+    (await import("./ts_route_mapper.ts")).default;
+
+  const compare: RouteComparator = opts?.compare ??
+    (await import("./pathname_lexical_route_comparator.ts")).default;
+
+  const readDir: DirectoryReader = opts?.readDir ??
+    (await import("./_read_dir.ts")).default;
 
   const routes: ComparableRoute[] = [];
 
   const iter = walk(
     !opts?.pattern || opts?.pattern === "/" ? "" : opts?.pattern ?? "",
     opts?.fileRootUrl ? fromFileUrl(opts.fileRootUrl) : ".",
+    readDir,
   );
 
   for await (const entry of iter) {
@@ -156,6 +109,9 @@ export async function discoverRoutes(
   routes.sort(compare);
 
   if (opts?.verbose) {
+    const { asSerializablePattern } = await import(
+      "./as_serializable_pattern.ts"
+    );
     for (const { pattern, module } of routes) {
       console.debug(
         `Found route: %c${asSerializablePattern(pattern)}%c -> %c${module}`,
@@ -172,14 +128,16 @@ export async function discoverRoutes(
 async function* walk(
   parentPattern: string,
   parentPath: string,
+  readDir: DirectoryReader,
 ): AsyncIterable<DiscoveredPath> {
-  for await (const entry of Deno.readDir(parentPath)) {
-    if (entry.isDirectory) {
+  for await (const entry of await readDir(parentPath)) {
+    if (isDirectory(entry)) {
       yield* walk(
         `${parentPattern}/${entry.name}`,
         join(parentPath, entry.name),
+        readDir,
       );
-    } else {
+    } else if (isFile(entry)) {
       const parsedName = parse(entry.name);
 
       yield {
@@ -190,56 +148,6 @@ async function* walk(
       };
     }
   }
-}
-
-/**
- * Default path mapper function.
- *
- * Adjust 'index' entry patterns to be the parent path with optional trailing slash.
- */
-export function defaultPathMapper(entry: DiscoveredPath): DiscoveredPath {
-  if (entry.name === "index") {
-    return {
-      ...entry,
-      pattern: entry.pattern === "/index"
-        ? "/"
-        : entry.pattern.replace(/\/index$/, "{/}?"),
-    };
-  }
-  return entry;
-}
-
-/**
- * Default route mapper function.
- *
- * Maps only TypeScript (ts/tsx) modules.
- */
-export function defaultRouteMapper(
-  { ext, pattern, module }: DiscoveredPath,
-): DiscoveredRoute[] {
-  switch (ext) {
-    case ".ts":
-    case ".tsx":
-      return [{
-        pattern,
-        module,
-      }];
-  }
-  return [];
-}
-
-/**
- * Default route comparator function.
- *
- * Really dumb string comparison of the pathname of the URLPattern.
- */
-export function defaultRouteCompare(
-  { pattern: { pathname: a } }: ComparableRoute,
-  { pattern: { pathname: b } }: ComparableRoute,
-): number {
-  // TODO: Sort using something like URLPattern.compareComponent if it becomes available
-  // See: https://github.com/WICG/urlpattern/issues/61
-  return a === b ? 0 : a > b ? -1 : 1;
 }
 
 /**
@@ -281,7 +189,7 @@ function consolidateRoutes(routes: ComparableRoute[]): DiscoveredRoute[] {
 }
 
 /**
- * Convert a module specifier to a URL object if can be parsed.
+ * Convert a module specifier to a URL object if it can be parsed.
  */
 function asModule(module: string | URL): string | URL {
   return module instanceof URL